Move ORMException inside the Exception namespace
