+++ lib/Doctrine/ORM/Exception/NotSupported.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Exception;
 
-use Doctrine\ORM\ORMException;
+use Doctrine\ORM\Exception\ORMException;
 
 final class NotSupported extends \Exception implements ORMException
 {
