+++ lib/Doctrine/ORM/Exception/RepositoryException.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Exception;
 
-use Doctrine\ORM\ORMException;
+use Doctrine\ORM\Exception\ORMException;
 
 /**
  * This interface should be implemented by all exceptions in the Repository
