+++ lib/Doctrine/ORM/Exception/SchemaToolException.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Exception;
 
-use Doctrine\ORM\ORMException;
+use Doctrine\ORM\Exception\ORMException;
 
 interface SchemaToolException extends ORMException
 {
