+++ lib/Doctrine/ORM/Mapping/Exception/TableGeneratorNotImplementedYet.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Mapping\Exception;
 
-use Doctrine\ORM\ORMException;
+use Doctrine\ORM\Exception\ORMException;
 
 final class TableGeneratorNotImplementedYet extends \Exception implements ORMException
 {
