+++ lib/Doctrine/ORM/Mapping/MappingException.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Mapping;
 
-use Doctrine\ORM\ORMException;
+use Doctrine\ORM\Exception\ORMException;
 use function array_keys;
 use function array_map;
 use function array_values;
