+++ lib/Doctrine/ORM/PessimisticLockException.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM;
 
+use Doctrine\ORM\Exception\ORMException;
+
 class PessimisticLockException extends \Exception implements ORMException
 {
     /**
