+++ lib/Doctrine/ORM/Query/QueryException.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Doctrine\ORM\Query;
 
 use Doctrine\ORM\Mapping\AssociationMetadata;
-use Doctrine\ORM\ORMException;
+use Doctrine\ORM\Exception\ORMException;
 use Doctrine\ORM\Query\AST\PathExpression;
 use function sprintf;
