+++ lib/Doctrine/ORM/Tools/Exception/MissingColumnException.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Tools\Exception;
 
-use Doctrine\ORM\ORMException;
+use Doctrine\ORM\Exception\ORMException;
 use function sprintf;
 
 final class MissingColumnException extends \Exception implements ORMException
