+++ lib/Doctrine/ORM/Utility/NormalizeIdentifier.php
@@ -8,7 +8,7 @@ use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Mapping\ClassMetadata;
 use Doctrine\ORM\Mapping\FieldMetadata;
 use Doctrine\ORM\Mapping\ToOneAssociationMetadata;
-use Doctrine\ORM\ORMException;
+use Doctrine\ORM\Exception\ORMException;
 use function array_key_exists;
 use function reset;
